Refactored a lot of Chris' code
